#42: move to Java 8 remove a LOT of warnings and update some dependencies

+++ hapi-hl7overhttp/src/main/java/ca/uhn/hl7v2/hoh/util/ByteUtils.java
@@ -15,7 +15,7 @@ public class ByteUtils {
 			byte nextByte = theBytes[i];
 			if (nextByte < ' ' || nextByte > 126) {
 				b.append('[');
-				b.append((int) nextByte);
+				b.append(nextByte);
 				b.append(']');
 			} else {
 				b.append((char) nextByte);
